# back-end/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "BED EP1 API - Martinus Nordgård",
    "description": "Documentation for the BED Exam Project API. Contains all nessecary information about how to use this API successfully. Documentation <del>automatically</del> partially generated by the swagger-autogen module and partially written by the Author."
  },
  "host": "localhost:3001",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/init": {
      "post": {
        "tags": [
          "Deployment and Maintenance"
        ],
        "summary": "Populate database ",
        "description": "This endpoint populated the database with Roles, Memberships, an Admin User, Products from the Noroff API and Order Statuses. This is achieved with running methods from various service files.",
        "responses": {
          "200": {
            "description": "Successful request",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Database tables have been populated successfully"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "statuscode": {
                  "type": "number",
                  "example": 500
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Error details"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/FAKEinit": {
      "post": {
        "tags": [
          "Deployment and Maintenance"
        ],
        "summary": "Populates database with local data. Used for testing.",
        "description": "Uses various services to populate the database, but doesn't not involve external API. Uses locally stored data to populate the products table",
        "responses": {
          "200": {
            "description": "Successful request",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "success"
                },
                "statuscode": {
                  "type": "number",
                  "example": 200
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Database tables have been populated successfully"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string",
                  "example": "error"
                },
                "statuscode": {
                  "type": "number",
                  "example": 500
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "string",
                      "example": "Error details"
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  }
}